refactor : Monster 관련 package 정리
