+++ src/main/kotlin/com/widehouse/dnd/challenge/AbilityCheck.kt
@@ -1,7 +1,7 @@
 package com.widehouse.dnd.challenge
 
-import com.widehouse.dnd.character.PlayerCharacter
 import com.widehouse.dnd.character.Skill
+import com.widehouse.dnd.character.player.PlayerCharacter
 import com.widehouse.dnd.dice.Dice
 
 class AbilityCheck(private val playerCharacter: PlayerCharacter, private val skill: Skill, private val difficultyClass: Int) {
