+++ src/main/kotlin/com/widehouse/dnd/character/nonplayer/MonsterSize.kt
@@ -1,0 +1,10 @@
+package com.widehouse.dnd.character.nonplayer
+
+enum class MonsterSize {
+    Tiny,
+    Small,
+    Medium,
+    Large,
+    Huge,
+    Gargantuan
+}
