+++ src/main/kotlin/com/widehouse/dnd/character/player/Race.kt
@@ -1,4 +1,4 @@
-package com.widehouse.dnd.character
+package com.widehouse.dnd.character.player
 
 enum class Race {
     Dwarf,
