+++ src/test/kotlin/com/widehouse/dnd/character/player/PlayerCharacterTest.kt
@@ -1,5 +1,9 @@
-package com.widehouse.dnd.character
+package com.widehouse.dnd.character.player
 
+import com.widehouse.dnd.character.Abilities
+import com.widehouse.dnd.character.AbilityType
+import com.widehouse.dnd.character.Constitution
+import com.widehouse.dnd.character.Strength
 import io.kotest.assertions.assertSoftly
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.collections.shouldContain
@@ -7,7 +11,15 @@ import io.kotest.matchers.shouldBe
 
 class PlayerCharacterTest : FreeSpec({
     "Player Character creation test" {
-        val char = PlayerCharacter.create(name = "foo", race = Race.Human, `class` = Class.Fighter, abilities = Abilities.of(15, 14, 13, 12, 10, 8))
+        val char = PlayerCharacter.create(name = "foo", race = Race.Human, `class` = Class.Fighter, abilities = Abilities.of(
+            15,
+            14,
+            13,
+            12,
+            10,
+            8
+        )
+        )
 
         assertSoftly(char) {
             name shouldBe "foo"
